libcamera: camera: Define an explicit pipeline model
Expand the existing implicit pipeline model with an explicit pipeline
API. The aim is to support advanced use cases where an application needs
to access frames between pipeline stages. In the Android world, this is
known as reprocessing.

This patch only contains the documentation that defines and details the
concepts and their usage. The rework of the camera configuration API
will follow.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>