libcamera: controls: Disable const move constructor and assignement
C++ supports const rvalue references. They are of dubious use, as the
purpose of rvalue references is to express move semantics, and moving a
const object is not possible, but they nonetheless exist. This means
that calling std::move() on a const reference will not produce an error,
but return a const rvalue reference.

Furthermore, a const lvalue reference can bind to a const rvalue. This
means that overload resolution will select an overload of a function
that takes a const lvalue reference (such as the copy constructor) when
passed a const rvalue reference (such as the return value of std::move()
called on a const reference) if no overload taking a const rvalue
reference exists.

The consequence of this is that the following code

	const ControlList &ctrls = ...;
	...
	ControlList copy(std::move(ctrls));

will select the ControlList copy constructor and thus compile, when one
may expect the object to be moved from the use of std::move() (there is,
on a side note, no guarantee in the language that a move constructor
will move anything).

To make the compiler reject usage of std::move() in this case, define
the copy and move constructor and assignment operators explicitly,
defaulting the non-const version of the move constructor and assignment
operator, but deleting the const version. The compiler will select the
const move version as it is declared, and generate an error as it is
deleted.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>