gstreamer: Add error checking in gst_libcamera_src_task_enter()
Error checking the output from generateConfiguration() was missing in
the code. Only assert was added as a guard which checked if the size of
generated camera config was equal to size of roles passed to it.

If the roles argument has a invalid member, it will return a nullptr and then
trying to access a member on a nullptr for size comparison will result in a
segmentation fault. So, if the function returns a nullptr, it will simply put
a error on GstBus and gracefully exit.

Signed-off-by: Vedant Paranjape <vedantparanjape160201@gmail.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>