android: Plumb lens focus distance
Plumb the controls related to lens focus distance:
- LENS_INFO_FOCUS_DISTANCE_CALIBRATION (static)
- LENS_INFO_MINIMUM_FOCUS_DISTANCE (static)
- LENS_FOCUS_DISTANCE (request, result)

Conceptually, the first two controls determine the range of possible
values of the third. None of our lenses are calibrated, so hardcode the
first to uncalibrated. The valid focus distances are then [0, min].
Since we don't yet support variable-focus lenses, hardcode the second to
0.0f, making the range of valid focus distances [0, 0]. Hence, hardcode
the focus distance in the result metadata to 0, and ignore the control
in the request. Add todos for adding proper controls later, once we have
focus support.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>