libcamera: base: thread: Prevent messages to thread without an event loop
A message posted to an object bound to a thread without an event loop
will never get delivered. This indicates a bug in the caller, and
materializes in ways that can be hard to debug, such as a signal never
being delivered. Add an assertion in Thread::postMessage() to catch this
at the source.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>