utils: ipu3-process: Stream multiple frames using v4l2-ctl
Stream IMGU using v4l2-ctl as it provides more granular controls
on the video nodes. The goal here is to stream a single input
file containing multiple frames in 10-bit IPU3 packed bayer format.
Such a input file can be created using the ipu3-pack utility.

This patch enables the IMGU's parameter node as well in
configure_pipeline() without which, the input node won't stream
successive input frames.

The viewfinder and output nodes will still stream using yavta, so
no changes required there.

It's is quite tricky to stream input and parameters nodes synchronously
with v4l2-ctl commands for each. The patch overcomes this short-coming
by streaming both the nodes endlessly until both of them are killed.
A sleep of 1 second is introduced to make sure $frame_count buffers
are captured on the output and vf nodes.

Another short-coming of this patch is when specific parameter-buffer(s)
(from a file) are needed to streamed on the parameters node, this might
require some adjusting. For now, the parameters node streams with
un-assigned buffers. It depends on the decision how this script is
expanded further down the line.

Signed-off-by: Umang Jain <umang.jain@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>