controls: Remove AeLocked
Now that the codebase has been ported to use the new AEGC controls
remove the definition of AeLocked. AeEnable is not be removed as it will
be redefined.

Signed-off-by: Jacopo Mondi <jacopo@jmondi.org>
Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Stefan Klug <stefan.klug@ideasonboard.com>

---
No change in v8

No change in v7

Changes in v6:
- remove only AeLocked, as AeEnable will be redefined instead of removed
  - (used to be "controls: Remove AeEnable and AeLocked")

No change in v5

No change in v4

No change in v3