libcamera: v4l2_device: Always set all four V4L2 colorspace fields
When setting a colorspace on a V4L2 video capture device or a subdev
source pad, the kernel considers the colorspace, xfer_func, ycbcr_enc
and quantization fields as four independent values. Any field set to
V4L2_*_DEFAULT by userspace is ignored, and the driver keeps its current
configuration for that field. This behaviour, while not necessarily well
known, is clearly documented in [1] and [2], and is implemented by the
rkisp1 driver.

The V4L2Device::fromColorSpace() function, when converting from a
libcamera ColorSpace to the V4L2 format colorspace-related fields, first
attempts to find a match for a colorspace preset. If found, it only sets
the colorspace field, and leaves the xfer_func, ycbcr_enc and
quantization fields to V4L2_*_DEFAULT. This contradicts the V4L2
specification, and prevents setting the transfer function, YCbCr
encoding and quantization on at least rkisp1.

Fix this issue by dropping the preset lookup, and set all four
colorspace-related fields explicitly.

[1] https://linuxtv.org/downloads/v4l-dvb-apis/userspace-api/v4l/pixfmt-v4l2.html#id2
[2] https://linuxtv.org/downloads/v4l-dvb-apis/userspace-api/v4l/subdev-formats.html#id2

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Tested-by: Jacopo Mondi <jacopo@jmondi.org>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>