py: examples: Add simple-capture-opencv.py
Add a simple example showing how the captured frame can be imported into
opencv and processed there.

This is only a simple example, and it has the same issue as all the
other examples: pixel format conversions. Each pixel format requires
custom conversion code, and I have implemented the conversion only for a
few of them.

Also, this is almost identical to simple-capture.py. Normally it would
make sense to add common helpers, but so far I have stayed away from it
to keep the examples independent and in a single file. Making the
examples "too fancy" will easily make them more difficult to understand
for a newcomer.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>