libcamera: utils: Wrap usage of lockf()
The lockf() function is not implemented in the Bionic standard C
library.

As an alternative, since Linux v2.0 the flock() function is available
as a direct system call instead of being emulated in the C library.

As lockf() is instead usually implemented as an interface to fcntl()
locking, locks placed by flock() and lockf() might not be detected.

As reported by the flock() function documentation:

Since kernel 2.0, flock() is implemented as a system call in its own right
rather than being emulated in the GNU C library as a call to fcntl(2).
With this implementation, there is no interaction between the types of
lock placed by  flock()  and  fc‐ ntl(2), and flock() does not detect
deadlock.  (Note, however, that on some systems, such as the modern
BSDs, flock() and fc‐ ntl(2) locks do interact with one another.)

To avoid risks of undetected deadlock, provide an wrapper function
utils, which in case flock() is not available, deflects all calls to
lockf() instead.

Signed-off-by: Jacopo Mondi <jacopo@jmondi.org>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>