libcamera: utils: Provide strchrnul() wrapper
The strchrnul function is a GNU-specific extension to string.h and it's
not availalable on all C libraries (in example, Android's Bionic).

Provide an implementation of the function based around the generally
available strchr() function and use it in the Logger class.

Open code the same wrapper in the application's option parser, where we
cannot use the internal utils namespace.

Signed-off-by: Jacopo Mondi <jacopo@jmondi.org>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>