py: cam: Network renderer
Here's something I have found useful a few times.

This adds a "tx" renderer to cam.py, which sends the frames over the
network to a receiver.

It also adds a "cam-rx" tool (non-libcamera based) which receives the
frames and uses PyQt to show them on the screen, usually ran on a PC.

This is obviously not super efficient, but on the PC side it doesn't
matter. On the TX side, at least RPi4 seemed to work without noticeable
lag, but my old 32-bit TI DRA76, when sending three camera streams, the
performance dropped to ~5fps. Still, I find that more than enough for
most development work.

This could be extended to also transmit the metadata.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>