libcamera: deprecated: Add new deprecated header
Provide a new header that will be included when applications
use <libcamera/libcamera.h> to bring in additional context and warnings
to generate when the libcamera API is modified where possible.

Introduce a deprecated implemenation of StreamRoles to report
that StreamRoles is no longer a libcamera definition.

This allows the compiler to report the following when the header is
available:

../../src/apps/qcam/main_window.cpp: In member function ‘int MainWindow::startCapture()’:
../../src/apps/qcam/main_window.cpp:366:21: error: ‘using StreamRoles = class std::vector<libcamera::StreamRole>’ is deprecated: Use a span, array or vector directly [-Werror=deprecated-declarations]
  366 |         StreamRoles roles = StreamKeyValueParser::roles(options_[OptStream]);
      |                     ^~~~~
In file included from include/libcamera/libcamera.h:16,
                 from ../../src/apps/qcam/main_window.cpp:14:
../../include/libcamera/deprecated.h:21:7: note: declared here
   21 | using StreamRoles [[deprecated("Use a span, array or vector directly")]]
      |       ^~~~~~~~~~~

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>