android: yuv: separate source destination in length check
Right now, the assumption of PostProcessorYuv is that both the source
and the destination are in formats::NV12.

This might change in the future, if we add other formats or pixel format
conversion as supported in libyuv.

Split out source and destination check to prepare for that.

No functional change.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>