android: yuv: loop over each plane for size check
PostProcessorYuv assumption is that only formats::NV12 is supported
for source and destination pixelFormat.

Because of this, we don't loop on each plane when size checking, but we
always assume a 2 plane buffer.

To prepare for adding more YUV formats such as YUYV, loop over the
planes instead of using an if.

No functional change, besides the logs only printing the first faulty
plane length.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>
Reviewed-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>