android: camera_device: support when no Direct stream is found
As the comment states, it's possible that the first element in
streamConfig.streams is a Internal stream.
In that case sourceStream stays a nullptr and all the Mapped streams are
mapped towards a nullptr.

Make sure that the Mapped streams are mapped to an Internal stream in case
that's the first element.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>