WIP: android: add YUYV->NV12 format conversion via libyuv
For some platforms, it's possible that the gralloc implementation
and the CSI receiver cannot agree on a pixel format. When that happens,
there is usually a m2m converter in the pipeline which handles pixel format
conversion.

On platforms without pixel format converters, such as the AM62x, we need to do
software conversion.

The AM62x platform:
* uses a CSI receiver (j721e-csi2rx), that only supports
  packed YUV422 formats such as YUYV, YVYU, UYVY and VYUY.
* Has a gralloc implementation that only supports of semi-planar
  YUV420 formats such as NV12.

Implement YUYV->NV12 format conversion using libyuv.

This is mainly done by transforming the first stream from Type::Direct into
Type::Internal so that it goes through the post-processor loop.

```
The WIP: part is mainly around computeYUYVSize():

Since gralloc and j721e-csi2rx are incompatible, we need a way to get
gralloc to allocate (NV12) the kernel-requested buffer length (YUYV).
In other words, we should make sure that the first plane of the NV12
allocated buffer is long enough to fit a YUYV image.

According to [1], NV12 has 8 bits (one byte) per component, and the
first plane is the Y component.
So a 1920x1080 image in NV12 has plane[0].length=1920*1080=2073600

According to [2], YUYV stores 2 pixels per container of 32 bits, which
gives us 16 bits (2 bytes for one pixel).
So a 1920x1080 image in YUYV has plane[0].length=1920*1080*2=4147200

So apply a *2 factor to make the kernel believe it's receiving a YUYV buffer.

Note: this also means that we are wasting NV12's plane[1] buffer with
each allocation.

[1] https://www.kernel.org/doc/html/latest/userspace-api/media/v4l/pixfmt-yuv-planar.html
[2] https://www.kernel.org/doc/html/latest/userspace-api/media/v4l/pixfmt-packed-yuv.html
```

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>