libcamera: rkisp1: Replace usage of RkISP1FrameInfo
Now that the pipeline handler can create a private derived class of
Request::Private, use it to store the pipeline specific data.

In the case of RkISP1 we associate the statistic and paramters buffers
with a Request.

As the IPA sends notifications for paramters and statistics buffer by
identifying them by frame id, associate the frame ids and the Request
in the RkISP1CameraData class.

This replaces the functionalities of RkISP1FrameInfo which can now be
removed.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Reviewed-by: Daniel Scally <dan.scally@ideasonboard.com>
Tested-by: Daniel Scally <dan.scally@ideasonboard.com>
Reviewed-by: Stefan Klug<stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>