libcamera: ipu3: Return Raw buffers on error
When a Request is cancelled, all its application-provided buffers
are completed and the paramters and stats buffer are made available
again to the pipeline handler.

If the RAW buffer doesn't come from the application (no RAW stream
requested) it comes from the CIO2Device internal pool. Return it to
the CIO2 pool in case of error.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Reviewed-by: Daniel Scally <dan.scally@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>