README.rst: Report py dependencies
The Python bindings were supported by a subproject when they were first
contributed to be able to directly make use of the smart pointers
branch.

This requirement has now been removed to prefer using released versions
of pybind11 directly, however the README.rst was not updated to
reference and reflect this.

Update the README.rst to report the pybind11 package which needs to be
installed to make use of the optional Python bindings component.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>