utils: libtuning: modules: Add skeletal AGC module
Add a skeletal AGC module just so that we can have some AGC tuning
values that we can use to test during development of AGC in the IPAs. As
rkisp1 is the main target, we only add support for rkisp1 for now.

The parameters are mostly copied from the hardcoded values in ctt,
except for the metering modes.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Stefan Klug <stefan.klug@ideasonboard.com>

---
Changes in v4:
- remove support for both v10 and v12, and only support v10
- remove piecewise linear function luminance targets

Changes in v3:
- remove unused compatible string
- make gain start from 2 in the exposure modes

Changes in v2:
- remove unneccessary imports
- add support for both v10 and v12