ipa: rkisp1: agc: Read histogram weights from tuning file
Add support to the rkisp1 AGC to read histogram weights from the tuning
file. As controls for selecting the metering mode are not yet supported,
for now hardcode the matrix metering mode, which is the same as what the
AGC previously hardcoded.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Stefan Klug <stefan.klug@ideasonboard.com>
Reviewed-by: Daniel Scally <dan.scally@ideasonboard.com>

---
Changes in v6:
- fix predivider calculation
- misc cleanups
- remove support for v12 (technically it depends on what the tuning file
  provides)

No change in v5

Changes in v4:
- add debug print when parsing the yaml file

Changes in v3:
- support the new tuning file layout for interpolated matrices
- support both v10 and v12

Changes in v2:
- add default metering mode if none are read successfully from the
  tuning file
- compute the predivider instead of using a table