libcamera: ipu3: Replace IPU3FrameInfo
Now that the pipeline handler can create a private derived class of
Request::Private, use it to store the pipeline specific data.

In the case of IPU3 we associate statistics and paramters buffer with a
Request and a raw buffer which gets dequeued from the CIO2 and input
to the ImgU input.

This replaces the functionalities of IPU3FrameInfo which can now be
removed.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>