pipeline: rkisp1: Apply initial controls
Controls passed to Camera::start() are not handled at the moment.
Implement that by initializing a separate frame context and then
returning the controls synchronously to the caller. In the pipeline the
controls get passed to the sensor before the call to streamOn() to
ensure the controls are applied right away. Passing the controls to the
pipeline using the setSensorControls signal is not appropriate because
it is asynchronous and would reach the sensor too late (initial controls
need to be applied before the sensor starts and before delayed controls
initializes its start condition).

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>

---

Changes in v3:
- reflowed commit message
- Dropped control list copy and unnecesary temporary