libcamera: lc-compliance: Add initial set of per-frame-control tests
These tests check if controls (only exposure time and analogue gain at
the moment) get applied on the frame they were requested for.

This is tested by looking at the metadata and the mean brightness
of the image center.

At the moment these tests fail. Fixes for the pipelines will be delivered
in later patches.

To run only the teste, one can run:
lc-compliance -c <cam> -f "SingleStream.*"

Note that the current implementation is a bit picky on what the camera
actually sees. If it is too dark (or too bright), the tests will fail.
Looking at a white wall in a normally lit office usually works.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>