libcamera: delayed_controls: Rework delayed controls implementation
Functional changes:
- Requests need to be queued for every frame
- The startup phase is no longer treated as special case
- Requests carry a sequence number, so that we can detect when the system
  gets out of sync
- Controls for a given sequence number can be updated multiple times
  as long as they are not yet sent out
- If it's too late, controls get delayed but they are not lost
- Requests attached to frame 0 don't get lost

Technical notes:
A sourceSequence_ replaces the updated flag to be able to track from which
frame the update stems. This is needed for the following use case:
Assume e.g. On frame 10 an ExposureTime=42 is queued because the user wants
manual exposure. Now the agc gets the stats for frame 8 (where auto
regulation is still active) and pushes a new ExposureTime for frame 9.
Frame 9 was already sent out, so it gets delayed to frame 11 (assuming a
delay of 2 on ExposureTime). This would revert the request from frame 10.
Taking the sourceSequence into account, the delayed request from frame
9 will be discarded, which is correct.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>