ipa: libipa: Add MatrixInterpolator class
Add a class to encapsulate the functionality of fetching a matrix based
on an integer key, and interpolating if there is no exact match. This is
expected to be used by both color correction matrices / crosstalk
correction as well as lens shading correction.

A cache is included only for exact matches of the key. The caller is
expected to decide the tolererance for rounding.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Stefan Klug <stefan.klug@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>

---
No change in v9

Changes in v8:
- use yaml get function
- use Matrix::identity() (instead of the default constructor, now that
  it makes a zero matrix instead of an identity matrix)

Changes in v7:
- remove cache
- fix copyright and license

Changes in v6:
- fix doxygen

Changes in v5:
- improve documentation
- replace count() with find() != end() (in get())
- add cache

Changes in v4:
- remove stray semicolons
- read from the new yaml layout (which mirrors what we have at the
  moment for lsc), and add keys to make it more flexible

Changes in v3:
- add a constructor that takes a map of unsigned int -> matrix
- s/unit/identity
- clear matrices on reset and when reading from yaml
- add assert at get()

Changes in v2:
- initialize to identity matrix
- add a function to reset to identity matrix
- other minor fixes