libcamera: lc-compliance: Add TimeSheet class
This class allows us to prepare a time sheet with controls for the frames
to capture (a bit like the capture script in cam).

During capture the metadata is recorded. Additionally a mean brightness
value of a 20x20 pixel spot in the middle of the frame is calculated.

This allows easy analysis after running the capture without complicated state
handling due to the asynchronous nature of the capturing process.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>