libcamera: lc-compliance: Add initial set of per-frame-control tests
Add tests that check if controls (only exposure time and analogue gain at the
moment) get applied on the frame they were requested for. This is tested by
looking at the metadata and the mean brightness of the image center. At the
moment these tests fail. Fixes for the pipelines will be delivered in later
patches (rkisp1 for now).

To run the pfc tests only:
lc-compliance -c <cam> -f "PerFrameControlTests.*"

Note that the current implementation is a bit picky on what the camera
actually sees. If it is too dark (or too bright), the tests will fail.
Looking at a white wall in a normally lit office usually works.

These tests are known to pass using a imx219 (RPi cam v2) with a imx8mp
(debix-som) using the rkisp1 pipeline

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>