libcamera: delayed_controls: Add sourceSequence property
This is necessary for the upcoming rework. It allows to track the
sequence number from which a request stems.

The use case where this is needed is as follows: Assume for frame 10 a manual
ExposureTime=42 is queued because the user wants manual exposure. Now the agc
gets the stats for frame 8 (where auto regulation is still active) and pushes a
new ExposureTime for frame 9.  Frame 9 was already sent out, so it gets delayed
to frame 11 (assuming a delay of 2 on ExposureTime). This would revert the
request from frame 10.  Taking the sourceSequence into account, the delayed
request from frame 9 will be discarded, which is correct.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>