pipeline: rkisp1: Apply initial controls
Controls passed to Camera::start() are now correctly applied.  This is needed
to pass the per-frame-controls test.  The code uses the return value of
ipa->start() to return the initial controls and pass them to the sensor instead
of the setSensorControls signal because setSensorControls is asynchronous and
would reach the sensor too late.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>