libcamera: device_enumerator: exclude uvcvideo device when enumerate
Camera service startup crash when camera and
usb camera are connected at the same time,
service will always reboot due to the null pointer
in populateEntities.

In this case, a UVC device was used to populate the
related entities, pads and links, which was not reasonable.

To fix the issue, ensure the media node is not
a "uvcvideo" device when enumerate before createDevice.

Change-Id: I656a542e83449df287d34a4e358ae2e26e784f17
Signed-off-by: Anle Pan <anle.pan@nxp.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>