libcamera: pipeline: Rename pipelines to a shorter name
The PipelineHandlerFactoryBase class has a name that is
propagated to the PipelineHandler instance it creates.

In present implementation, this name comes from the
REGISTER_PIPELINE_HANDLER registration macro. It corresponds to
the stringified name of the PipelineHandler derived class.
Therefore, PipelineHandler factories and instances names can be
quite long such as "PipelineHandlerRkISP1".

A libcamera user may have to explicitly refer to a PipelineHandler
name for configuration purpose: one usage of the name can be to
define a pipeline handlers match list and their priorities.
It is desired, for user convenience, to use a short name to
designate a pipeline handler. Reusing the short pipeline names
already defined in the meson option files is an existing and
consistent way of naming pipelines.

This change adds an explicit name parameter to the
REGISTER_PIPELINE_HANDLER registration macro. That parameter is
used to define the name of a pipeline handler factory, instead of
the current pipeline handler class name.

Each pipeline registration is updated accordingly. The short name
assigned corresponds to the pipeline directory name in the
source tree. It is consistent with pipelines names used in meson.

Changing the pipeline name has an impact on the IPA modules:
each module defines a IPAModuleInfo structure. This structure has
a pipelineName member defining the pipeline handler name it shall
match with.
Therefore, each internal IPA module definition has to be changed
to have its IPAModuleInfo pipelineName name updated with the short
pipeline handler name.

In addition to this pipelineName member, the IPAModuleInfo structure
also has a name member, associated to the IPA module name.
Having renamed the pipelines to a short name, the pipeline name and
the IPA module names of the IPAModuleInfo structure are the same:
for in-tree IPA, they correspond to the respective pipeline and IPA
subdirectories in the source tree.
However the IPA name could be different, for instance with a close
source IPA implementation built out-of-tree. Thus, it makes sense to
keep the IPA name in that structure, as the 2 definitions may not
always be redundant.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>